requirements.txt
python-telegram-bot>=22.7